- Module Refactoring - Unused code removed - Code improvements - Minor bug removal
