+++ core/database/src/main/java/com/jet/database/JetDatabase.kt
@@ -4,7 +4,7 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.example.restaurant_impl.database.entities.Restaurant
 import com.jet.database.dao.RestaurantDao
-import com.jet.database.data.entities.RestaurantStatus
+import com.jet.database.entities.RestaurantStatus
 
 @Database(entities = [Restaurant::class, RestaurantStatus::class], version = 1)
 abstract class JetDatabase: RoomDatabase() {
