+++ core/database/src/main/java/com/jet/database/dao/RestaurantDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
 import com.example.restaurant_impl.database.entities.Restaurant
-import com.jet.database.data.entities.RestaurantStatus
+import com.jet.database.entities.RestaurantStatus
 import kotlinx.coroutines.flow.Flow
 
 @Dao
@@ -18,7 +18,7 @@ interface RestaurantDao {
         "SELECT * FROM restaurant_table " +
                 "JOIN restaurant_status_table ON restaurant_table.status = restaurant_status_table.status " +
                 "WHERE name LIKE '%' || :query || '%' ORDER BY " +
-                "CASE WHEN :sortOrder = 'status' THEN restaurant_status_table.value END ASC," +
+                "restaurant_status_table.value ASC," +
                 "CASE WHEN :sortOrder = 'distance' THEN distance END ASC," +
                 "CASE WHEN :sortOrder = 'bestMatch' THEN best_match END DESC," +
                 "CASE WHEN :sortOrder = 'newest' THEN newest END DESC," +
@@ -27,12 +27,9 @@ interface RestaurantDao {
                 "CASE WHEN :sortOrder = 'averageProductPrice' THEN average_product_price END ASC," +
                 "CASE WHEN :sortOrder = 'deliveryCosts' THEN average_product_price END ASC," +
                 "CASE WHEN :sortOrder = 'minCost' THEN min_cost END ASC"
-    )
-    suspend fun getRestaurantList(query: String, sortOrder: String): List<Restaurant>
 
-    @Transaction
-    @Query("SELECT * FROM restaurant_table")
-    fun getRestaurantListFlow(): Flow<List<Restaurant>>
+    )
+    fun getRestaurantList(query: String, sortOrder: String): Flow<List<Restaurant>>
 
     @Insert
     suspend fun insertRestaurantStatus(status: List<RestaurantStatus>)
