+++ feature-restaurant/src/main/AndroidManifest.xml
