+++ feature-restaurant/src/main/assets/restaurants.json
