+++ feature-restaurant/src/main/java/com/example/restaurant_impl/NavigationCommand.kt
