+++ feature-restaurant/src/main/java/com/example/restaurant_impl/domain/RestaurantRepository.kt
@@ -1,0 +1,8 @@
+package com.example.restaurant_impl.domain
+
+import com.example.restaurant_impl.database.entities.Restaurant
+import kotlinx.coroutines.flow.Flow
+
+interface RestaurantRepository {
+    suspend fun getRestaurantsList(query: String, sortValue: String): Flow<List<Restaurant>>
+}
