+++ feature-restaurant/src/main/java/com/example/restaurant_impl/ui/viewstates/RestaurantListViewState.kt
