+++ feature-restaurant/src/main/res/values/strings.xml
