--- feature/restaurant/restaurant-impl/src/main/java/com/example/restaurant_impl/RestaurantModule.kt
@@ -1,18 +1,0 @@
-package com.example.restaurant_impl
-
-import com.example.restaurant_api.data.RestaurantRepository
-import com.example.restaurant_impl.data.RestaurantRepositoryImpl
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object RestaurantModule {
-
-    @Singleton
-    @Provides
-    fun provideRepository(repository: RestaurantRepositoryImpl): RestaurantRepository = repository
-}
